refactor: SRP 원칙에 따른 클래스 분리에 맞춰 테스트 클래스도 분리

+++ src/test/java/com/nexus/sion/feature/squad/query/service/SquadQueryServiceImplTest.java
@@ -71,7 +71,7 @@ class SquadQueryServiceImplTest {
             .domainCount(3)
             .weight(0.8)
             .monthlyUnitPrice(500000)
-            .productivity(BigDecimal.valueOf(1.2))
+            .productivity(1.2)
             .build();
 
     given(squadQueryMapper.findJobsByProjectId(projectId)).willReturn(List.of(job1));
